+++ NEWS.md
@@ -3,6 +3,7 @@
 * GitHub actions are now used for continuous integration instead of Travis and
 AppVeyor.
 * Added precomp function to avoid repeated computation.
+* mrs_data2mat function now collapses all dimensions to dynamics by default.
 
 # spant 1.7.0
 * Added write_mrs function which guesses the output format from the file